Introduce `UV.Internal.Util`, remove `UV.Loop.defaultLoop`
Introduce `UV.Internal.Utils`, a foreign imported data type that gives
access to common used functions, such as `Left`, `Right`, `Just, and
`Nothing`. We can keep this utility bundle in static storage and share
it betwen usages. Further, we can partially apply it to those
functions that need it, further reducing it's associated overhead.

Coincidentally, this commit also removes `UV.Loop.defaultLoop`.  This
change resulted from moving the `UV.Internal.Utils` injection from the
time of running a uv loop to the time of it's creation. In order for
the default loop to function properly (concretely, communicate results
via `Left`, and `Right`) we need to occupy the `uv_loop_t->data`
field. This implies mutating the default loop if we gave access to it.
The alternative would be to require an library initialization call
which would prime the utilities in static storage.